% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idw.R
\name{.grid_over}
\alias{.grid_over}
\title{Regular grid over a set of points}
\usage{
.grid_over(x, y, n = 80, expand = 0.05)
}
\arguments{
\item{x, y}{Point coordinates.}

\item{n}{Cells along the longer axis.}

\item{expand}{Fraction of the range to pad on each side.}
}
\value{
A data frame of grid \code{x}, \code{y}.
}
\description{
Regular grid over a set of points
}
\keyword{internal}
